refactor(SectionDao): 인터페이스로 추상화
