+++ src/main/java/wooteco/subway/dao/SectionDao.java
@@ -1,0 +1,10 @@
+package wooteco.subway.dao;
+
+import java.util.List;
+import wooteco.subway.domain.Section;
+
+public interface SectionDao {
+    Section save(Long lineId, Section section);
+
+    List<Section> findByLineId(long lineId);
+}
